added the geomagic description package in the geomagic touch repo. intergrated urdf model from francesco to the controller of dane.
renamed some phantom omni stuff to geomagic (not all)

added a few parameters in the launch file
