+++ geomagic_control/launch/geomagic_headless.launch
@@ -1,0 +1,24 @@
+<launch>
+
+   <arg name="device_name" default="Geomagic"/>
+   <param name="publish_rate" type="int" value="100" />
+   
+   
+   
+	<!-- omni node publishes joint states on omni1_joint_states -->
+        <node name="$(arg device_name)" pkg="geomagic_control" type="omni" output="screen">
+            <param name="device_name" type="str" value="$(arg device_name)"/>
+            
+        </node>
+	
+	<!-- robot_description tells robot_state_publisher where to find the omni's urdf description -->
+        <param name="$(arg device_name)_robot_description" command="cat $(find geomagic_description)/urdf/omni.urdf" />
+
+	<!-- robot_state_publisher reads Geomagic_joint_states and publishes corresponding tfs -->
+        <node name="$(arg device_name)_robot_state_publisher" pkg="robot_state_publisher" type="robot_state_publisher">
+	  <remap from="joint_states" to="$(arg device_name)_joint_states" />
+	  <remap from="robot_description" to="$(arg device_name)_robot_description" />
+	</node>
+	
+</launch>
+
