+++ geomagic_control/msg/PhantomButtonEvent.msg
@@ -1,0 +1,2 @@
+int32 grey_button
+int32 white_button
