+++ geomagic_description/CMakeLists.txt
@@ -1,0 +1,6 @@
+cmake_minimum_required(VERSION 2.8.3)
+project(geomagic_description)
+
+find_package(catkin REQUIRED COMPONENTS urdf)
+
+catkin_package()
