+++ geomagic_description/urdf/omni.robot.xml
@@ -1,0 +1,6 @@
+<Robot name="omni" file="omni.dae">
+  <Manipulator name="arm">
+    <base>base</base>
+    <effector>stylus</effector>
+  </Manipulator>
+</Robot>
