--- msg/PhantomButtonEvent.msg
@@ -1,2 +1,0 @@
-int32 grey_button
-int32 white_button
