--- rosdep.yaml
@@ -1,27 +1,0 @@
-x11proto-dri2-dev:
-    ubuntu: x11proto-dri2-dev
-libdrm-dev:
-    ubuntu: libdrm-dev
-x11proto-gl-dev:
-    ubuntu: x11proto-gl-dev
-libx11-dev:
-    ubuntu: libx11-dev
-libexpat1-dev:
-    ubuntu: libexpat1-dev
-libxt-dev:
-    ubuntu: libxt-dev
-libmotif-dev:
-    ubuntu: libmotif-dev
-x11proto-print-dev:
-    ubuntu: x11proto-print-dev
-freeglut3-dev:
-    ubuntu: freeglut3-dev
-tcsh:
-    ubuntu: tcsh
-libncurses5-dev:
-    ubuntu: libncurses5-dev
-libgl1-mesa-dev:
-    ubuntu: libgl1-mesa-dev
-libglw1-mesa-dev:
-    ubuntu: libglw1-mesa-dev
-
